Add Wavefront metrics and trace exporters

+++ src/Management/src/EndpointBase/Metrics/IMetricsEndpointOptions.cs
@@ -6,6 +6,6 @@ namespace Steeltoe.Management.Endpoint.Metrics
 {
     public interface IMetricsEndpointOptions : IEndpointOptions
     {
-        long ScrapeResponseCacheDurationMilliseconds { get; }
+        int ScrapeResponseCacheDurationMilliseconds { get; }
     }
 }
